refactor product model, price data type

+++ backend/DTOs/Product/AddProductDTO.cs
@@ -6,7 +6,7 @@ namespace backend.DTOs.Product;
 public class AddProductDTO: BaseModel
 {
     public string? Name { get; set; }
-    public float? Price { get; set; }
+    public decimal? Price { get; set; }
     public string? Description { get; set; }
     public string? Image { get; set; }
     public int CategoryId { get; set; }
