+++ backend/DTOs/Product/UpdateProductDTO.cs
@@ -9,7 +9,7 @@ public class UpdateProductDTO
 {
     public int Id { get; set; }
     public string? Name { get; set; }
-    public float? Price { get; set; }
+    public decimal? Price { get; set; }
     public string? Description { get; set; }
     public string? Image { get; set; }
     public int CategoryId { get; set; }
